Rename yfc500 folder to more common YardForce
