+++ Firmware/CoverUI/platformio.ini
@@ -40,7 +40,7 @@ build_src_filter =
   -<LEDcontrol.cpp>
   -<statemachine.cpp>
   -<build/*>
-  -<yfc500/stm32cube/*>
+  -<YardForce/stm32cube/*>
 
 build_flags =
   -DHW_YFC500
@@ -59,14 +59,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f0x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
   -c
   reset_config none separate
 
@@ -81,14 +81,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f0x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
   -c
   reset_config none separate
 
@@ -104,14 +104,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f0x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
   -c
   reset_config none separate
 
@@ -126,14 +126,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f0x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-stm.cfg
+  YardForce/util/stlink-stm.cfg
   -c
   reset_config none separate
 
@@ -154,7 +154,7 @@ platform_packages = platformio/tool-openocd@^2.1100.211028
 upload_protocol = cmsis-dap
 upload_flags =
   -f
-  yfc500/util/picoprobe-stm.cfg
+  YardForce/util/picoprobe-stm.cfg
 
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f0x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
@@ -164,7 +164,7 @@ debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/picoprobe-stm.cfg
+  YardForce/util/picoprobe-stm.cfg
   -c
   reset_config none separate
 
@@ -186,14 +186,14 @@ build_flags =
   
 upload_flags =
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f3x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
   -c
   reset_config none separate
 
@@ -212,14 +212,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f3x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
   -c
   reset_config none separate
 
@@ -239,14 +239,14 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f3x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
   -c
   reset_config none separate
 
@@ -265,13 +265,13 @@ build_flags =
 
 upload_flags =
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
 upload_command = openocd $UPLOAD_FLAGS -c "init; reset halt; stm32f3x unlock 0; reset run" -c "program {$SOURCE} verify reset; shutdown;"
 
 debug_build_flags = -O0 -g -ggdb2
 debug_server = 
   openocd
   -f
-  yfc500/util/stlink-gd.cfg
+  YardForce/util/stlink-gd.cfg
   -c
   reset_config none separate
